Create Neo4j-Sveriges Regering
Create a Neo4j graph database containing:

Node types
By node type we mean a node with a certain label:

Sweden's government ministers/state councilors (see https://www.regeringen.se/sveriges-regering) - name, title/ministerial role, year of birth (see respective CV), motto, party affiliation
The different departments - name and description
The different "areas" - name and description
Consider: Can party affiliation be a label? Pros and cons of this?

Relationships
Which ministers belong to which department?
What areas is a certain minister responsible for?
Which areas is a certain department responsible for?
Check your database
Find all ministers belonging to a particular party.
Find all ministers belonging to a certain department.
Find all areas for which a certain department is responsible.
Find all the areas a particular minister is responsible for.
Find which department is responsible for a certain area.
Find which minister is responsible for a certain area.Ministers' ages - which department is governed by the youngest and oldest ministers?
If you are missing any ministers' birth years, try to find them!
Ask/look up
Produce a list of the departments and the average age of each department's ministers.
Arrange the list from lowest to highest average age.
Decades
Create a node for each decade (from the decade the oldest minister was born in, to the decade the youngest politician was born in.). Each decade should have an attribute name (eg '1970s'), start year (eg 1970) and end year (eg 1979), as well as - if you want - a brief description of the decade.
Create relationships between the ministers and the decades. The relationship is directed (“born under”).
Questions/search
Can you show which politicians were born in a certain decade?
During which decade were most of the L ministers born? Most of the KD ministers? Most of the M ministers.
To think about: Which decades are least/most similar to each other (culturally? politically?). Was it possible to create relationships in the style of "most similar" between decades?
The governing parties and the left-right scale
If you haven't already (eg if you only have parties as labels or properties on ministers) create a node for each government party (with party name and party abbreviation as properties).
Create relationships between the parties "politically closest" so that L and KD are tied together and KD and M are tied together. This type of relationship is not directed. (Although many consider the left-right scale in politics to be a blunt measure, for the parties in government they are usually arranged from left to right like this: L - KD - M.)
Then create relationships that bind each minister to their party (“is a member of”). This type of relationship is directed.
Questions/search
How do you search/display:
All ministers within a certain party.
The distance between different lots (0 to 2).
Can you also list the political distance between different ministers based on the distance between parties? How?
Is it possible to arrange the department on the left-right scale based on which parties their ministers belong to?
« Thomas homework for Friday March 8 - done!